:lipstick: Updated loaders & responsiveness
