+++ components/Album/Tags.js
@@ -2,8 +2,11 @@ import { Tag, Link } from "@chakra-ui/react";
 import React from "react";
 import NextLink from "next/link";
 
-function Tags({ tag }) {
-  console.log(tag);
+function Tags(tag) {
+  console.log(tag.tag);
+
+  const tagName = tag.tag[0];
+  const tagURL = tag.tag[1];
   return (
     <div>
       <Tag
@@ -19,12 +22,12 @@ function Tags({ tag }) {
         maxW="120px"
       >
         <Link
-          href={tag.url}
+          href={tagURL}
           textDecoration="none"
           _hover={{ color: "purple.600" }}
         >
           {" "}
-          {tag.name}
+          {tagName}
         </Link>
       </Tag>
     </div>
