+++ components/Artist/Artist.js
