+++ pages/home.js
@@ -12,49 +12,21 @@ import {
   Flex,
 } from "@chakra-ui/react";
 import axios from "axios";
-import Album from "../components/Album";
+import Album from "../components/Album/Album";
 import { tagTopAlbumsFetch } from "../utils/fetch";
 import useSWR from "swr";
 import fetcher from "../utils/fetcher";
 import { useAuth } from "../providers/authProvider";
-import TrendingArtists from "../components/TrendingArtists";
+import TrendingArtists from "../components/Artist/TrendingArtists";
 import TrendingTracks from "../components/Tracks/TrendingTracks";
+import TopAlbums from "../components/Album/TopAlbums";
 
 function home() {
   const auth = useAuth();
 
-  //   const [randomNum, setNum] = useState(0);
-  //   const [albums, setAlbums] = useState([]);
   let albums = [];
   const emojis = ["👋", "🌊", "🔊", "💿", "🎸", "🥁", "🎶", "🎺", "🎻", "🎤"];
   let randomNum = Math.floor(Math.random() * 9);
-
-  const { data, error, isValidating } = useSWR(
-    tagTopAlbumsFetch + "&tag=dance",
-    {
-      revalidateOnFocus: false,
-      refreshWhenOffline: false,
-      refreshWhenHidden: false,
-      refreshInterval: 0,
-      dedupingInterval: 1000000,
-    }
-  );
-
-  if (error || isValidating) {
-    console.log("poop");
-  } else {
-    albums = data.albums.album;
-    // setAlbums(data.albums.album);
-
-    // albums.map((album) => {
-    //   console.log(album);
-    // });
-  }
-
-  //   useEffect(() => {
-  //     // fetchAlbums();
-  //     setNum(Math.floor(Math.random() * 9));
-  //   }, [randomNum]);
 
   return (
     <>
@@ -86,41 +58,7 @@ function home() {
           <Heading>top dance albums</Heading>
           <Text>all time</Text>
         </Box>
-        {error || isValidating ? (
-          <Center>
-            <SimpleGrid columns="1" row="2" gap={1}>
-              <Box>
-                <Skeleton
-                  startColor="purple.500"
-                  endColor="orange.500"
-                  height="25rem"
-                  width="100%"
-                />
-              </Box>
-              <Box>
-                <Skeleton
-                  startColor="orange.500"
-                  endColor="purple.500"
-                  height="25rem"
-                  width="100%"
-                />
-              </Box>
-            </SimpleGrid>
-          </Center>
-        ) : (
-          <Grid
-            gridTemplateColumns={[
-              "repeat(2, 1fr)",
-              "repeat(2, 1fr)",
-              "repeat(5, 1fr)",
-            ]}
-            gap={3}
-          >
-            {albums.map((album) => (
-              <Album key={album.artist.mbid} thing={album} />
-            ))}
-          </Grid>
-        )}
+        <TopAlbums />
         <Box mt={10}>
           <Box mb={4}>
             <Heading>trending artists</Heading>
